Changed ParamsModel. Completed minimal frontend. Reworked Injector, Extractor, Optimizer algorithms

+++ app/optimizer_api/database/models/params.py
@@ -13,15 +13,15 @@ class ParamsModel(Base):
 
     param_number = Column(Integer)
 
-    freq_bottom = Column(Numeric, nullable=False)
-    freq_top = Column(Numeric, nullable=False)
-    duration = Column(Numeric, nullable=False)
+    freq_bottom = Column(Integer, nullable=False)
+    freq_top = Column(Integer, nullable=False)
+    duration = Column(Integer, nullable=False)
     freq_bottom_grad = Column(Numeric)
     freq_top_grad = Column(Numeric)
     duration_grad = Column(Numeric)
-    freq_bottom_step = Column(Numeric, nullable=False)
-    freq_top_step = Column(Numeric, nullable=False)
-    duration_step = Column(Numeric, nullable=False)
+    freq_bottom_step = Column(Integer, nullable=False)
+    freq_top_step = Column(Integer, nullable=False)
+    duration_step = Column(Integer, nullable=False)
 
     expert_score = Column(Numeric)
     sound_noise_ratio = Column(Numeric)
